Move {HierarchyView,petgraph} from views/hierarchy into views/
