+++ src/hugr/views.rs
@@ -1,12 +1,13 @@
 //! Read-only access into HUGR graphs and subgraphs.
 
 pub mod hierarchy;
+mod petgraph;
 pub mod sibling_subgraph;
 
 #[cfg(test)]
 mod tests;
 
-pub use hierarchy::{DescendantsGraph, HierarchyView, SiblingGraph};
+pub use hierarchy::{DescendantsGraph, SiblingGraph};
 pub use sibling_subgraph::SiblingSubgraph;
 
 use context_iterators::{ContextIterator, IntoContextIterator, MapWithCtx};
@@ -18,8 +19,8 @@ use super::{Hugr, NodeMetadata, NodeType};
 use crate::ops::handle::NodeHandle;
 use crate::ops::{FuncDecl, FuncDefn, OpName, OpTag, OpType, DFG};
 use crate::types::{EdgeKind, FunctionType};
-use crate::Direction;
-use crate::{Node, Port};
+use crate::{Direction, Node, Port};
+use ::petgraph::visit as pv;
 
 /// A trait for inspecting HUGRs.
 /// For end users we intend this to be superseded by region-specific APIs.
@@ -226,6 +227,27 @@ pub trait HugrView: sealed::HugrInternals {
             })
             .finish()
     }
+}
+
+/// A common trait for views of a HUGR hierarchical subgraph.
+pub trait HierarchyView<'a>:
+    HugrView
+    + pv::GraphBase<NodeId = Node>
+    + pv::GraphProp
+    + pv::NodeCount
+    + pv::NodeIndexable
+    + pv::EdgeCount
+    + pv::Visitable
+    + pv::GetAdjacencyMatrix
+    + pv::Visitable
+where
+    for<'g> &'g Self: pv::IntoNeighborsDirected + pv::IntoNodeIdentifiers,
+{
+    /// The base from which the subgraph is derived.
+    type Base;
+
+    /// Create a hierarchical view of a HUGR given a root node.
+    fn new(hugr: &'a Self::Base, root: Node) -> Self;
 }
 
 impl<T> HugrView for T
